requirements.txt
Django==3.2.12
djangorestframework==3.13.1
djangorestframework_simplejwt==5.1.0
djongo==1.3.6
django_cors_headers==3.11.0
dnspython==2.2.1
pymongo==3.12.3
python-decouple==3.6
Pillow==9.0.1
channels ==3.5.0
channels-redis==3.2.0
wheel==0.37.1
celery==5.2.3
django-celery-results==2.3.0
django-celery-beat==2.2.1

websocket-client==0.57.0

